refactor(skills): rename 11 legacy skills to prompt.md + skill.config.ts
Brings the 11 legacy skills (skill.md + config.ts) onto the same naming
as the 5 newer skills (triage, repo-match, bug-enhance, echo-test,
echo-test-holdout). Also updates relative ./config.js imports inside
each renamed skill's slice.test.ts.

Static @goose-hub/skills/.../config.js imports and the read-prompt.ts
loader still point at the old names — those are fixed in the next
commits.

Part of the skill convention consolidation playbook
(plans/skill-convention-consolidation.md, step 4.1).
