+++ skills/advise-on-plan/prompt.md
