+++ skills/advise-on-plan/skill.config.ts
