+++ skills/investigate/prompt.md
