+++ skills/investigate/skill.config.ts
