+++ skills/playwright-repro/slice.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { toJSONSchema } from 'zod';
-import config, { PlaywrightReproContextSchema } from './config.js';
+import config, { PlaywrightReproContextSchema } from './skill.config.js';
 import { PlaywrightReproSchema } from './schema.js';
 
 describe('playwright-repro schema', () => {
