+++ skills/qa/slice.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { toJSONSchema } from 'zod';
-import config, { QaContextSchema } from './config.js';
+import config, { QaContextSchema } from './skill.config.js';
 import {
   CriteriaResultSchema,
   FindingSchema,
