+++ skills/resolve-conflict/prompt.md
