+++ skills/resolve-conflict/skill.config.ts
