+++ skills/retrospective-light/slice.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { toJSONSchema } from 'zod';
-import config from './config.js';
+import config from './skill.config.js';
 import { LightRetroSchema } from './schema.js';
 
 const baseValid = {
