+++ skills/spec-author/prompt.md
