+++ skills/spec-author/skill.config.ts
